Fixing timing.
L1 T1 would be 256, and I was clipping at 255. This threw the math off and made things go out of sync. Added quick workaround to properly handle 256.

+++ Sequencer.h
@@ -117,8 +117,8 @@ bool sequencerStart();
 bool sequencerStop();
 bool sequencerIsPlaying();
 bool sequencerIsReady();
-bool sequencerPut(byte track, byte note, byte noteLength);
-bool sequencerGet(byte track, byte *note, byte *noteLength);
+bool sequencerPut(byte track, byte note, unsigned int noteLength);
+bool sequencerGet(byte track, byte *note, unsigned int *noteLength);
 bool sequencerHandler();
 
 unsigned int sequencerBufferAvailable();
